[Packet] change unsigned long to size_t for windows

+++ src/common/network/DataPacket.cpp
@@ -18,7 +18,7 @@ babel::common::DataPacket::getCommandId() const
 	return _commandId;
 }
 
-unsigned long babel::common::DataPacket::getNbArgs() const
+size_t babel::common::DataPacket::getNbArgs() const
 {
 	return _args.size();
 }
